bonus-app/src/main/java/com/example/bonusapp/service/CustomerService.java
package com.example.bonusapp.service;

import com.example.bonusapp.entity.Customer;
import com.example.bonusapp.repository.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class CustomerService {

    private final CustomerRepository customerRepository;

    @Autowired
    public CustomerService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

   public Customer createOrSaveCustomer(Customer customer) {
         return customerRepository.save(customer);
   }

    public Optional<Customer> getCustomerById(Long customerId) {
        return customerRepository.findById(customerId);
    }
}
